Initial development

+++ controllers/admin.js
@@ -1,14 +1,17 @@
 const mongodb = require('mongodb');
 
 const Product = require('../model/product');
+const User = require('../model/user');
+const path = require('path');
+const fs = require('fs');
 
-const ObjectId = mongodb.ObjectId;
+
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
-    pageTitle: 'Add Product',
-    path: '/admin/add-product',
-    editing: false
+    // pageTitle: 'Add Product',
+    // path: '/admin/add-product',
+    // editing: false
   });
 };
 
@@ -16,20 +19,41 @@ exports.postAddProduct= (req, res, next) => {
   const title = req.body.title;
   const price = req.body.price;
   const description = req.body.description;
-  const imageUrl = req.body.imageUrl;
-  const creator = req.body.creator;
+  const image = req.file;
+  console.log(image);
+  let creator;
+  if(!image){
+    const error = new Error('Inavalid file attach');
+    error.statusCode = 422;
+    throw error;
+  }
+  
+  const imageUrl = image.path;
+  console.log(imageUrl);
   //console.log(req.body);
   const product = new Product ({
       title: title, 
       price: price, 
       description: description, 
       imageUrl: imageUrl,
-      creator: creator
+      creator: req.userId
     });
-  product.save()
+  product
+  .save()
   .then(result => {
+    return User.findById(req.userId);})
+    .then(user => {
+      creator= user;
+      user.products.push(product);
+      return user.save();
+    })
+    .then( result => {
     console.log('created product');
-    res.status(201).json({ message: 'Product created!', productId: result._id });
+    res.status(201).json({ message: 'Product created!', productId: result._id,
+    product: product,
+    creator: {_id: creator._id, name: creator.name}
+  });
+    
   })
   .catch(err => {
     next(err) ;
@@ -37,37 +61,63 @@ exports.postAddProduct= (req, res, next) => {
   };
 
   exports.getEditProduct = (req, res, next) => {
-    const editMode = req.query.edit;
-    console.log(editMode);
-    if (!editMode) {
-      return res.redirect('/');
-    }
+    // const editMode = req.query.edit;
+    // console.log(editMode);
+    // if (!editMode) {
+    //   return res.redirect('/');
+    // }
     const prodId = req.params.productId;
     Product.findById(prodId)
       .then(product => {
         if (!product) {
-          return res.redirect('/');
+          const error = new Error('Could not find product.');
+          error.statusCode = 404;
+          throw error;
+          //return res.redirect('/');
         }
         console.log(product);
-        res.status(201).json({ message: 'This is edit product', productId: product._id });
+        res.status(201).json({ message: 'This is edit product', product: product });
       })
-      .catch(err => console.log(err));
+      .catch(err => 
+        next(err));
   };
   
 
   exports.postEditProduct = (req, res, next) => {
-    const prodId = req.body.prodId;
+    const prodId = req.params.productId;
+    //console.log(prodId);
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    let imageUrl = req.body.image;
+    if (req.file) {
+      imageUrl = req.file.path;
+    }
+    if (!imageUrl) {
+      const error = new Error('No file picked.');
+      error.statusCode = 422;
+      throw error;
+    }
     const updatedDesc = req.body.description;
     
     Product.findById(prodId)
       .then(product => {
+        if (!product) {
+          const error = new Error('Could not find product.');
+          error.statusCode = 404;
+          throw error;
+        }
+        if (product.creator.toString() !== req.userId) {
+          const error = new Error('Not authorized!');
+          error.statusCode = 403;
+          throw error;
+        }
+        if (imageUrl !== product.imageUrl) {
+          clearImage(product.imageUrl);
+        }
         product.title = updatedTitle;
         product.price = updatedPrice;
         product.description = updatedDesc;
-        product.imageUrl = updatedImageUrl;
+        product.imageUrl = imageUrl;
         return product.save();
       })
       .then(result => {
@@ -80,7 +130,18 @@ exports.postAddProduct= (req, res, next) => {
   };
   
   exports.getProducts = (req, res, next) => {
+    const currentPage = req.query.page || 1;
+    const perPage = 1000;
+    let totalItems;
     Product.find()
+      .countDocuments()
+      .then(count => {
+        totalItems = count;
+        console.log(totalItems);
+        return Product.find()
+          .skip((currentPage - 1) * perPage)
+          .limit(perPage);
+      })
       .then(products => {
         console.log(products);
         res.status(201).json({ message: 'Show all products', products: products });
@@ -88,8 +149,21 @@ exports.postAddProduct= (req, res, next) => {
         
       })
       .catch(err => {
-        console.log(err);
+        if (!err.statusCode) {
+          err.statusCode = 500;
+        }
+        next(err);
       });
+    // Product.find()
+    //   .then(products => {
+    //     console.log(products);
+    //     res.status(201).json({ message: 'Show all products', products: products });
+    //     return products;
+        
+    //   })
+    //   .catch(err => {
+    //     next(err);
+    //   });
   };
 
   
@@ -101,13 +175,44 @@ exports.postAddProduct= (req, res, next) => {
     
   
   exports.postDeleteProduct = (req, res, next) => {
-    const prodId = req.body.productId;
+    const prodId = req.params.productId;
+    console.log(prodId);
     //const prodId = '62e4f15b6feac670df19fd73';
-    Product.findByIdAndRemove(prodId)
+    Product.findById(prodId)
+      .then(product => {
+        if(!product){
+          const error = new Error('Could not find product');
+          error.statusCode = 404;
+          throw error;
+        }
+        if (product.creator.toString() !== req.userId){
+          const error = new Error ('Not Authorized');
+          error.statusCode = 403;
+          throw error;
+        }
+        clearImage(product.imageUrl);
+        return Product.findByIdAndRemove(prodId);
+      })
+      .then(result => {
+        return User.findById(req.userId);
+      })
+      .then(user => {
+        user.products.pull(prodId);
+        return user.save();
+      })
       .then(() => {
         console.log('DESTROYED PRODUCT');
         res.status(201).json({ message: 'Deleted product'});
         
       })
-      .catch(err => console.log(err));
+      .catch(err => 
+        next(err));
+  };
+
+
+  const clearImage = filePath => {
+    filePath = path.join(__dirname, '..', filePath);
+    fs.unlink(filePath, 
+      err => 
+      console.log(err));
   };
